+++ Fun.Blazor/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Add more helpers for html.renderFragment
+- Improve script tag
 
 ## [3.3.0-beta006] - 2023-12-11